:recycle: Rename loginV2 to login and update references on code
