+++ src/components/app/index.test.js
@@ -8,7 +8,7 @@ import { I18nextProvider } from 'react-i18next';
 import i18n from '../../i18n'; // initialized i18next instance
 import App from '.';
 import Splashscreen from '../splashscreen/splashscreen';
-import Login from '../loginV2/loginV2';
+import Login from '../login/login';
 import TransactionsDashboard from '../transactionDashboard';
 import routes from '../../constants/routes';
