+++ src/components/signInTooltipWrapper/signInTooltipWrapper.js
@@ -16,7 +16,7 @@ const SignInTooltipWrapper = ({
         content={React.cloneElement(children, { className: `${children.props.className} ${styles.child} disabled` })}
         title={t('Please sign in')}
         footer={(
-          <Link to={`${routes.loginV2.path}?referrer=${pathname}${encodeURIComponent(search)}`}>
+          <Link to={`${routes.login.path}?referrer=${pathname}${encodeURIComponent(search)}`}>
             {t('Sign in')}
           </Link>
 )}
